MoviePresenter creating cache module and moving MovieImageCache to it
